finishing the main function
Modifying main in so that instead of outputting the number of letters, words, and sentences, it instead outputs (only) the grade level as defined by the Coleman-Liau index (e.g. "Grade 2" or "Grade 8" or the like).
If the resulting index number is 16 or higher (equivalent to or greater than a senior undergraduate reading level), the program will output "Grade 16+" instead of outputting an exact index number. If the index number is less than 1, the program will output "Before Grade 1".